feat: 유저에 비밀번호 필드 추가 (Lv3 완성)

+++ src/main/java/com/example/schedulerv2/service/UserService.java
@@ -1,6 +1,9 @@
 package com.example.schedulerv2.service;
 
-import com.example.schedulerv2.dto.*;
+import com.example.schedulerv2.dto.user.UserCreateRequest;
+import com.example.schedulerv2.dto.user.UserCreateResponse;
+import com.example.schedulerv2.dto.user.UserGetResponse;
+import com.example.schedulerv2.dto.user.UserUpdateResponse;
 import com.example.schedulerv2.entity.User;
 import com.example.schedulerv2.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -22,7 +25,8 @@ public class UserService {
         // 유저 객체 만들어서 정보 저장
         User user = new User(
                 userCreateRequest.getUsername(),
-                userCreateRequest.getEmail()
+                userCreateRequest.getEmail(),
+                userCreateRequest.getPassword()
         );
 
         // DB에 유저 정보 저장
@@ -48,13 +52,21 @@ public class UserService {
 
     // 유저 닉네임 수정
     @Transactional
-    public UserUpdateResponse updateUser(Long id, String username) {
+    public UserUpdateResponse updateUser(Long id, String username, String password) {
 
         // ID 값 여부 체크
         checkId(id);
 
         // 유저 여부 체크
         User user = userRepository.findUserByIdOrElseThrow(id);
+
+        // 비밀번호 체크
+        checkPassword(password);
+
+        // 비밀번호가 비교
+        if (!user.getPassword().equals(password)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
 
         // 유저 닉네임 null, 공백 체크
         if (username == null || username.isBlank()) {
@@ -73,13 +85,21 @@ public class UserService {
 
     // 유저 삭제
     @Transactional
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id, String password) {
 
         // ID 값 여부 체크
         checkId(id);
 
         // 유저 여부 체크
         User user = userRepository.findUserByIdOrElseThrow(id);
+
+        // 비밀번호 체크
+        checkPassword(password);
+
+        // 비밀번호가 비교
+        if (!user.getPassword().equals(password)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
 
         // DB에 유저 정보 삭제
         userRepository.delete(user);
@@ -95,6 +115,12 @@ public class UserService {
     // ID 값 여부 체크
     private void checkId(Long id) {
         if (id == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID는 필수 입력값입니다");
+        }
+    }
+
+    private void checkPassword(String password) {
+        if (password == null || password.isBlank()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID는 필수 입력값입니다");
         }
     }
